requirements-web.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0 